Add command hooks mechanism for simple dependency resolution.
The following hooks have been added:

1. `before_add_command:<command>` - Called before WP_CLI tries to add the command `<command>` to its known vocabulary.

The callback that is being called by this hook will receive an instance of a `WP_CLI\Dispatcher\CommandAddition` object, which allows the callback to control whether WP_CLI should proceed with the addition or not. On aborting, the callback can provide a reason.

This can be used to let the addition of a command depend on some prior requirement to be met.

2. `after_add_command:<command>` - Called after the command `<command>` has been added to the known vocabulary.

This can be used for let the addition of the command depend on other commands being added first. In this way, the addition of commands can be rendered independent of loading order.

Note: Introducing the class `WP_CLI\Dispatcher\CommandAddition` was necessary in order to get around the fact that arguments cannot be reliably passed across the flow `do_hook()` -> `call_user_func_array()` -> `add_hook()`.

Fixes #3835